Dockerizing CDMLoader and adding Maven wrapper
Dockerizing done.
The Dockerfile expects a CDM.tar.gz to be present under 1-read-manifest/code-java project directory
The CDM.tar.gz contains the entire zipped up CDM project.
In the future , it should be pruned to contain only the CDM corpus files - after it is clear what are those files that CDM project depends on.

The CDM.tar.gz is not in repository due to its size;
Please look in docker_commands.txt in /docs to see how it is to be created first.
Currently the path in docker i.e. /CDM/samples is hardcoded;
It should be changed to an environment variable for configuration purposes

The CDMLoader docker container takes a SEARCH_SERVICE env variable which tells where Weaviate is. This is used to load the CDM into Weaviate.
See docker_commands.txt for an example